checkprocfs: check procfs
When deploying tetragon in kind (or similar environments) the resulting
procfs, will not be the host procfs. This can cause issues. This patch
detects this situation and issues a warning so that it can be
identified.

It retrieves the host pid via a simple program, and checks it against
the self link in the procfs directly.

In a kind deployment, this produces:
> time="2023-09-01T16:02:33Z" level=warning msg="pid mismatch: Is procfs the host /proc?" host-pid=1406461 proc-pid=4364 procfs=/procRoot

If a VM (e.g., with minikube --driver=qemu2) is used, then no warning appears.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>